[WIP] LTTng-Analyses integration
TODO:
- Split up, first commit should be just empty plugins (+ test plugin)
- Merge rest of wip-lami-analysis into this

This patch introduces a new "analysis.lami" plugin, which contains
generic support for any analysis implementing the LAMI protocol [1],
as well as integration with the current LTTng-Analyses.

If LTTng-Analyses are present on the user's system (and on the $PATH),
options under the "External Analyses" trace sub-tree element become
available. Right-clicking on those allows running the specified
analysis, and will import the results as Reports in Trace Compass.

Reports are opened automatically when created, and will initially
display a table of the results returned by the external analysis.
The toolbar menu can then be used to generate charts from this data,
either Histograms or XY scatter plots.

Clicking on chart elemnents highlights the corresponding column in the
table, and vice versa. Also, if the data table rows contain a time
range or timestamp, integration via standard time selection signal
is done.

To not commit to final APIs yet, we've used the "internal.provisional"
naming, as described in [2].

Full user documentation and additional unit tests will follow.

Current limitations:
- The RunAnalysisHandler and others are specific to LAMI analyses,
  which   means other types of OndemandAnalyses would have to define
  their own way of being started (which may not even be through
  a right-click menu option).

[1] https://github.com/lttng/lttng-analyses/blob/master/doc/mi.md
[2] https://wiki.eclipse.org/Provisional_API_Guidelines

Change-Id: I3cee70f20347ed2c71f077d40117073aeaf4161d
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>
Signed-off-by: Philippe Proulx <pproulx@efficios.com>