ss: Implement state aggregation rules
State aggregation can be used to define states without
doing explicit state insertions, like with modifyAttribute().

It is now possible to "mount" aggregation rules to some
attributes of the state system (think mounting a device
to a path on the filesystem). Queries to these attributes
will use the aggregation rule to get the corresponding state.

This can be helpful in modeling complex attributes by modeling
simple, individual states separately and aggregating them into
one big attribute that views/analyses can use directly.

This initial implementation provides 3 aggregation rules,
an AttributePriority, a BitwiseOr and a SymbolicLink one.

The first could be useful for modeling kernel CPU states, and
the second  SoftIRQ states, similarly to what is shown in the
unit tests.
The third one is just a generic helper that could be used
in many places.

Change-Id: I347c6c6419c2fe0d30e31543cd45070e5a00c505
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>