tmf: Make Callstack View pass timestamps and PIDs to symbol provider
Now that the Callstack analysis and view track the PIDs of each
function entry/exit, it can pass them along with the timesteamps
to the symbol providers, so that provider implementations making
use of the new methods can resolve symbols in the view too.

Also add proper comparison logic for ProcessEntries, they can use
numerical comparaison of the PID they now carry.

Change-Id: I7404ab4b893878dbdd8e526a646014739cfbe228
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>