lttng.ust: Split the function name into its own aspect/column
Introduce a new aspect for the Function Location. For now it will
be used to store the function name, but could eventually also
print the offset within the function once we can retrieve it.

Implement a new Callsite object for this analysis which will not
print the function name, since we will have it separately.

Change-Id: Ie7d1598a2bfebe690c3e82183e8ad16f62489b4e
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>