tmf: Return correct type in SourceLookupPropertySource
The "Line" value of the Source Lookup event property displays
correctly, but throws an AssertionFailedException upon clicking
on it.

This is due to the property source returning a Long object
instead of a String. It was using Long.valueOf(long) instead
of String.valueOf(long). Changing the return type highlights
it clearly.

Change-Id: Ie12813c67ef754b2bcf1bb38804b08565847dea8
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>