analysis.lami: Implementation of LAMI plugins
If pre-defined scripts are present on the user's system (and on the
$PATH), options under the "External Analyses" trace sub-tree element
become available. Right-clicking on those allows running the specified
analysis, and will import the results as Reports in Trace Compass.

Reports are opened automatically when created, and will initially
display a table of the results returned by the external analysis.
The toolbar menu can then be used to generate charts from this data,
either Histograms or XY scatter plots.

Clicking on chart elemnents highlights the corresponding column in the
table, and vice versa. Also, if the data table rows contain a time
range or timestamp, integration via standard time selection signal
is done.

To not commit to final APIs yet, we've used the "internal.provisional"
naming, as described in [1].

Full user documentation and additional unit tests will follow.

Current limitations:
- The RunAnalysisHandler and others are specific to LAMI analyses,
  which means other types of OnDemandAnalyses would have to define
  their own way of being started (which may not even be through
  a right-click menu option).

[1] https://wiki.eclipse.org/Provisional_API_Guidelines

Change-Id: I188116fa2c0c9a485aabd89d49cae0d6e73136b8
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>
Signed-off-by: Philippe Proulx <pproulx@efficios.com>