analysis.lami: Add support for user-defined LAMI analyses
This patch adds support for adding user-defined LAMI analyses,
as well as removing them, through the UI (project explorer).

A .properties file is created in the workspace for each added
user-defined analysis.

To add a new external analysis, right-click the External Analyses
item in the Project Explorer View, and click Add External Analysis.

To remove an external analysis, right-click the item to remove,
and click Remove External Analysis.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>