analysis.lami: Move built-in LTTng analyses to config files
This patch moves the knowledge of "shipped" LTTng analyses to
individual configuration files so that it is possible to add/remove
shipped LAMI analyses without having to edit any Java file.

The `index.properties` file contains a single property, `analyses`,
which is a whitespace-separated list of names of configuration files,
without the extension, to load in order to create as many LAMI
analyses.

Each configuration file contains two properties:

  * `name`: Display name of the analysis.
  * `command`: Full command to run. This string supports
    double-quoted arguments to allow arguments with whitespaces,
    or empty arguments. The `\` and `"` characters must be
    escaped with `\`.

Both properties are mandatory.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>