Selection and graph beautification
On table viewer creation with a time range set the current selection that
intersect with the time range. In most case if there was a selected time range
all elements present in the viewers will be in selected state.

Use color fading for the scatter chart and draw a cross for each selected item.

On mouse hovering for scatter chart draw a cross that intersect with both axis.

Also the radius for selection and hovering cross snapping is increased.

Change-Id: I48d2a258832c302b21e0579695f4ff83196e12c1